remove unnecessary examples methods
