+++ annotation_pipeline/utils.py
@@ -1,0 +1,9 @@
+import ibm_boto3
+from ibm_botocore.client import Config
+
+
+def get_ibm_cos_client(config):
+    return ibm_boto3.client(service_name='s3',
+                            ibm_api_key_id=config['ibm_cos']['api_key'],
+                            config=Config(signature_version='oauth'),
+                            endpoint_url=config['ibm_cos']['endpoint'])
